Rename harness to agGridReact
